Start of Arduino game. Added basic arrow key controls for debugging until the Arduino is hooked up, although right now only one key can be pressed at once resulting in jerky movement but this won't be a problem when hooked up.
Next, there needs to be some sort of map with a basic collision system. The idea being, to make the most cut down version of labyrinth possible to begin with so the potentiometers will directly control the position of the ball(square atm), to be later converted to using gravity at a later date with a pivoting platform.

The player right not consists of just a SDL_Rect object but I am considering making the player its own class to keep movement and lives more organised.

As far as the Arduino C code, I need to find where the last version of it is which was set up to send information from two potentiometers, which is all that's needed for the prototype.